Add mono build script (#70) with Travis CI and add config file
This add limited support for Travis CI. At the moment, we only build
the R Provider and do not run the tests. This is future work - we need
to figure out how to install R on Travis first.

The commit also includes some work on adding configuration file
(which is completed in a later commit).