Add  type provider for reading RData files (close #59, replace #85)
This implements `RData` type provider that can be used for reading RData
files in a statically typed way. The provider infers the types of the fields
by using the default conversion (if there is one available). More details
are available in the newly added documentation file.

Aside from the RData provider, I also updated F# Formatting (and the default
templates), added generated API reference (and tweaked a few things so that
it looks nice) and also updated F# Data reference to keep things up-to-date.