Fix failing unit test (conversion of strings via factors)
I was a bit surprised by this, because the bug must have been around for a while
(we need to add continuous integration for R provider and run the tests automatically!)

The problem is that the IntegerVector in recent versions of R.NET only
succeeds when the value is *not* a factor (there is an explicit check
`not (sexp.IsFactor())`) and so the pattern was never matched in the
convertor. This change uses R.NET only to check that the value is *some*
vector and then explicitly checks for numeric vectors (without ruling out
factors).