Register InteropServer as a well-known singleton
Issue #86 is caused by the server object being gc'd at the server. This
can be prevented by registering the server object as a
WellKnownServiceTypeEntry and specifying that the service be provided as
a singleton. This guarantees that the lifetime of the server object will
be the lifetime of the server process.