Improve exception forwarding for RInteropServer
When an exception is thrown during the execution of a method in
RInteropServer that exception cannot be returned to the client unless it
is serializable. All methods now wrap their behavior in a try statement
that catches all exceptions and grabs the message from non-serializable
exceptions to re-raise as a serializable exception.