Release fix for F# 3.1.2 (shadow copying)
This also changes R.NET reference to require a specific version. This
is more correct, because otherwise the `RProvider.fsx` script won't work
when a new version of R.NET is released (it has a specific version in
the `#r` path). The same is a problem for Deedle and FsLab (fix follows).