Use ILRepack to make RProvider.Server.exe completely stand-alone
As suggested by @forki, this change uses ILRepack to merge FSharp.Core.dll (and RProvider.Runtime.dll)
with RProvider.Server.exe. This means that the executable is standalone and we do not have to worry
about mismatching versions of FSharp.Core (or NuGet/VS getting confused because there is FSharp.Core
in the RProvider package - that's no longer needed).