Update R provider to use Paket & try fix Travis
The only notable thing about this is that I had to move RProvider.DesignTime to a
separate directory (a good thing anyway) so that it can have its own `paket.references`
file (specifying additional dependencies). This is now also using github reference
to get the type provider starter pack files.