Expose Server via an interface and remove dead code
The server needs to be exposed as an interface, because F# Interactive on Mono
doesn't seem to like DLL files referencing EXE files (so the client cannot
reference the server directly).

This also removes the (never used) option for local execution, because this would
need testing (if we need it, it can be easily recovered) & one not used function.
Added documentation as a bonus :-).