Make base functionality available without open RProvider.``base`` (fix #2)
I implemented this while doing other changes and forgot about it :-).
This is done simply by providing another "R" type, directly inside
the RProvider namespace with a copy of all "base" functions (if you
open both, you'll get the more recently opened one, which is fine as
they're the same).