Workaround #140
As of today Cecil (used in ILRepack) fails to link generic parameters
in some (nested generics types+methods) scenarios.
A fix was attempting in later versions but it fails in other scenarios.
Version 1.22.2 is still the most stable one, and it manages to merge
your deps (didn't check further than the build success though).